libsensors: use int64_t instead of long long int
Timestamps are of the type int64_t in sensors_event_t structure.

+++ libsensors/lsm330dlc_acceleration.c
@@ -47,7 +47,7 @@ void *lsm330dlc_acceleration_thread(void *thread_data)
 	struct input_event event;
 	struct timeval time;
 	struct lsm330dlc_acc acceleration_data;
-	long long int before, after;
+	int64_t before, after;
 	int diff;
 	int device_fd;
 	int uinput_fd;
